+++ Assets/Scripts/EnemyStats.cs
@@ -4,7 +4,7 @@ using System;
 public class EnemyData
 {
     public EnemyStats enemyStats;
-    public float hp;
+    public int hp;
     public float baseDmg;
     public bool alive; 
 }
@@ -13,7 +13,7 @@ public class EnemyStats : ScriptableObject
 {
     public string enemyName;
 
-    public float maxHp;
+    public int maxHp;
     public Sprite combatSprite;
     public Skill skill1;
     public Skill skill2;
